refactor: Remove lodash dep; add util.js
lodash is definitely useful, but it is **over 5MB**.  That is a lot, considering it was only used for type checking and assigning defaults.

Defaults are now handled by the `Object.assign` builtin.  Some type checking is now performed by a new `util.js` module, and some is performed inline with builtins and expressions.

And now the whole package is 5MB lighter!  That’s one hell of a diet. ☻